pyproject.toml
[tool.poetry]
name = "board"
version = "0.1.0"
description = ""
authors = ["ksg9482 <ksg930523@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
aiosqlite = "^0.20.0"
black = "^24.4.2"
pre-commit = "^3.7.1"
isort = "^5.13.2"
pycln = "^2.4.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
sqlmodel = "^0.0.19"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
pytest-asyncio = "^0.23.8"
aiomysql = "^0.2.0"
cryptography = "^43.0.0"
pydantic-settings = "^2.4.0"
locust = "^2.31.3"
pytest-env = "^1.1.3"
python-ulid = "^2.7.0"
mypy = "^1.11.2"
types-passlib = "^1.7.7.20240819"
pydantic = "^2.8.2"
pytest-mock = "^3.14.0"
testcontainers = {extras = ["mysql"], version = "^4.8.0"}
redis = "^5.0.8"
google-cloud-storage = "^2.18.2"
apscheduler = "^3.10.4"
email-validator = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.pycln]
all = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
env = [
  "DATABASE_URL=mysql+aiomysql://user:password@127.0.0.1:3306/testexampledb",
  "TESTING=True"
]
testpaths = [
  "tests"
]
markers = [
  "create",
  "patch",
  "put",
  "delete",
  "posts",
  "post",
  "comments",
  "comment",
  "login",
  "logout",
  "unit",
  "get"
]